+++ src/app/blog/page.jsx
@@ -1,22 +1,20 @@
-"use client";
-
 import styles from "@/app/blog/blogPage.module.css";
 import CardList from "@/components/cardList/CardList";
 import CategoryList from "@/components/categoryList/CategoryList";
 import Menu from "@/components/menu/Menu";
-import { useSearchParams } from "next/navigation";
 
-const BlogPage = () => {
-  const searchParams = useSearchParams();
-  const page = Number(searchParams.get("page")) || 1;
-  const cat = searchParams.get("cat") || "";
+const BlogPage = async ({ searchParams }) => {
+  const params = await searchParams;
+  const page = Number(params.page || "1");
+  const cat = params.cat || "";
+  const tags = params.tags || "";
 
   return (
     <div className={styles.container}>
       <h1 className={styles.title}>{cat ? `${cat}` : "Blog"}</h1>
       <CategoryList />
       <div className={styles.content}>
-        <CardList page={page} cat={cat} />
+        <CardList page={page} cat={cat} tags={tags} />
         <Menu />
       </div>
     </div>
