+++ src/app/tags/page.jsx
@@ -1,42 +1,23 @@
-"use client";
-import TagList from "@/components/tagList/TagList";
 import styles from "./tagsPage.module.css";
 import CardList from "@/components/cardList/CardList";
-import { useSearchParams } from "next/navigation";
-import { useRouter } from "next/navigation";
+import dynamic from "next/dynamic";
 
-const TagsPage = () => {
-  const searchParams = useSearchParams();
-  const page = Number(searchParams.get("page")) || 1;
-  const router = useRouter();
-  const rawTags = searchParams.get("tags") || "";
+const TagsPageClient = dynamic(() => import("./tagsPageClient"), {
+  loading: () => <p>loading...</p>,
+});
+
+const TagsPage = async ({ searchParams }) => {
+  const params = await searchParams;
+  const page = Number(params.page) || 1;
+  const rawTags = params.tags || "";
   const tags = rawTags ? rawTags.split(".").filter((tag) => tag !== "") : [];
 
-  const handleTagClick = (tagName) => {
-    const params = new URLSearchParams(searchParams.toString());
-    const currentTags = params.get("tags") ? params.get("tags").split(".") : [];
-    if (!currentTags.includes(tagName)) {
-      currentTags.push(tagName);
-      params.set("tags", currentTags.join("."));
-    } else {
-      // 태그 다시 누르면 제거
-      const filtered = currentTags.filter((t) => t !== tagName);
-      if (filtered.length > 0) {
-        params.set("tags", filtered.join("."));
-      } else {
-        params.delete("tags");
-      }
-    }
-    params.set("page", "1");
-
-    router.push(`/tags?${params.toString()}`);
-  };
   return (
     <div className={styles.container}>
       <h1></h1>
-      <TagList selectedTags={tags} onTagClick={handleTagClick} />
+      <TagsPageClient initialTags={tags} initialPage={page} />
       <div className={styles.content}>
-        <CardList />
+        <CardList page={page} tags={tags} />
       </div>
     </div>
   );
