+++ src/components/featured/Featured.jsx
@@ -1,3 +1,4 @@
+"use client";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import React from "react";
