+++ src/components/menuTags/MenuTags.jsx
@@ -1,3 +1,4 @@
+"use client";
 import Link from "next/link";
 import React, { useEffect, useState } from "react";
 import styles from "./menuTags.module.css";
